requirements.txt
django
python-dotenv
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-ckeditor
django-taggit
dj-taggit-serializer
pillow